{"schema":"decodex/commit/1","summary":"Split MCP server core modules","authority":"XY-1115"}
